requirements.txt
endee
sentence-transformers
PyPDF2
streamlit
groq
python-dotenv